Moved sysinfo to models
